Update texasholdem.py
I know ya all are probably not gonna read this until you all wake up, and to be fair, it's still a work in progress, but that's what I have so far for texas holdem poker. Sorry that it took me so long to contribute for code as well and not just UI and design.